+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 <a name="0.0.0"></a>
-# 0.0.0 (2018-08-03)
+# 0.0.0 (2018-08-06)
 
 
 ### Bug Fixes
 
 * **Business Application[Issue [305](https://chaos.ns.team/nsv2/ns-ops/issues/305) - business-application-validation-issue-on-documentation-step]:** added or validation on doucmentation step for tax return section ([b9212ae](https://chaos.ns.team/nsv2/ns-client/commits/b9212ae))
 * **Business Application[Issue [309](https://chaos.ns.team/nsv2/ns-ops/issues/309) - business-application-issues-finding-fix]:** business application finding fixes ([10c2ce1](https://chaos.ns.team/nsv2/ns-client/commits/10c2ce1))
+* **Business Application[Issue [309](https://chaos.ns.team/nsv2/ns-ops/issues/309) - business-application-issues-finding-fix]:** file uploader added confirm modal on remove file operation ([a99c1e4](https://chaos.ns.team/nsv2/ns-client/commits/a99c1e4))
 * **UserManagement[Issue [263](https://chaos.ns.team/nsv2/ns-ops/issues/263) - user management ui fixes from review]:** user management ui fixes from review ([b47f55e](https://chaos.ns.team/nsv2/ns-client/commits/b47f55e))
 
 
@@ -19,6 +20,7 @@
 * **Hello World structure [Issue [204](https://chaos.ns.team/nsv2/ns-ops/issues/204) - Setup Hello World structure for Issuer/Admin]:** Hello World structure for Issuer/Admin ([76810e0](https://chaos.ns.team/nsv2/ns-client/commits/76810e0))
 * **home page ui[Issue [228](https://chaos.ns.team/nsv2/ns-ops/issues/228) - public pages home page 2]:** created ui of home page ([e0e48e7](https://chaos.ns.team/nsv2/ns-client/commits/e0e48e7))
 * **Investor Account Details[Issue [289](https://chaos.ns.team/nsv2/ns-ops/issues/289) - view transactions and settings for investor account details]:** view transactions and settings for investor account details ([f2a77e7](https://chaos.ns.team/nsv2/ns-client/commits/f2a77e7))
+* **Investor summary dashboard[Issue [320](https://chaos.ns.team/nsv2/ns-ops/issues/320) - investor summary after account setup]:** investor summary after account setup ([a5a426c](https://chaos.ns.team/nsv2/ns-client/commits/a5a426c))
 * **investor-account-details-overview-ui[Issue [301](https://chaos.ns.team/nsv2/ns-ops/issues/301) - 301 investor account details overview ui]:** Added overview page UI on investor account details page ([6acae58](https://chaos.ns.team/nsv2/ns-client/commits/6acae58))
 * **Investor[Issue [221](https://chaos.ns.team/nsv2/ns-ops/issues/221) - investor signup flow ui ux updates 2]:** Review Feedback points - spelling correction and IRA - funding default set to null ([28904b1](https://chaos.ns.team/nsv2/ns-client/commits/28904b1))
 * **Investor[Issue [235](https://chaos.ns.team/nsv2/ns-ops/issues/235) - manage-beneficiaries-feedback-changes]:** investor beneficiary feedback changes ([9718a27](https://chaos.ns.team/nsv2/ns-client/commits/9718a27))